requirements.txt
streamlit
openai>=1.3.0
requests
beautifulsoup4
python-dotenv
pandas
browser-use
playwright
langchain>=0.1.15 
langchain-openai>=0.0.7
spacy
tiktoken 
chromadb 
faiss-cpu 
duckduckgo-search
langchain-community
pypdf
langsmith
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
